# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Roadmap for Jira API",
        "description": "Endpoint mapping for Roadmap for Jira App.",
        "contact": {
            "email": "contact@mrrafael.ca"
        },
        "version": "2.0"
    },
    "servers": [
        {
            "url": "http://localhost:3030",
            "description": "Server"
        }
    ],
    "paths": {
...
        "/api/v1/project/config": {
            "get": {
                "tags": [
                    "Project"
                ],
                "description": "Load a List of Projcts config objects.",
                "operationId": "getProjects",
                "responses": {
                    "200": {
                        "description": "Project config list loaded successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "JSON object containing a list of Projects configs",
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/ProjectConfigModelList"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "JSON object containing the error message",
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/ErrorModel"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Project"
                ],
                "description": "Save the project param.",
                "operationId": "saveProject",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProjectConfigModel"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Project config saved.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "JSON object containing the saved project config",
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/ProjectConfigModel"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "JSON object containing the error message",
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/ErrorModel"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Project"
                ],
                "description": "Remove the project param.",
                "operationId": "removeProject",
                "requestBody": {
                    "description": "Project Key",
                    "required": true,
                    "content": {
                        "text/plain": {
                            "schema": {
                                "type": "string",
                                "example": "ABC"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "JSON object containing the error message",
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/ErrorModel"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ProjectConfigModel": {
                "required": [
                    "projectKey",
                    "startDate",
                    "dueDate"
                ],
                "properties": {
                    "projectKey": {
                        "type": "string",
                        "example": "ABC"
                    },
                    "startDate": {
                        "type": "string",
                        "example": "customfield_123"
                    },
                    "dueDate": {
                        "type": "string",
                        "example": "customfield_456"
                    }
                },
                "type": "object"
            },
            "ProjectConfigModelList": {
                "required": [
                    "list"
                ],
                "properties": {
                    "list": {
                        "type": "array",
                        "items": {
                            "oneOf": [
                                {
                                    "$ref": "#/components/schemas/ProjectConfigModel"
                                }
                            ]
                        }
                    }
                },
                "type": "object"
            },
...
            "ErrorModel": {
                "required": [
                    "code",
                    "error"
                ],
                "properties": {
                    "code": {
                        "type": "string",
                        "example": "42S22"
                    },
                    "error": {
                        "type": "string",
                        "example": "Error message"
                    },
                    "message": {
                        "type": "string",
                        "example": "Complementary Error message"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
}